+++ bigml/tests/test_14_create_evaluations.py
@@ -125,7 +125,7 @@ class TestEvaluation(object):
         """
         print self.test_scenario3.__doc__
         examples = [
-            ['data/iris.csv', '50', '50', '800', '80', 'average_phi', '0.94107']]
+            ['data/iris.csv', '50', '50', '800', '80', 'average_phi', '0.89054']]
         for example in examples:
             print "\nTesting with:\n", example
             source_create.i_upload_a_file(self, example[0])